+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Ce projet adhère au principe du [Semantic Versioning](https://semver.org/spec/v
 ## 0.12.0 (unreleased)
 
 - Améliore le calcul du matériel restant dans les événements.
+- Ajoute la possibilité de limiter les caractéristiques spéciales du matériel par catégorie (#91).
 
 ## 0.11.0 (2021-01-14)